# api/openapi.yaml
openapi: 3.1.0
info:
  title: MaxPosty API
  version: 1.0.0
  description: |
    Canonical HTTP contract shared by the MaxPosty backend and frontend.
    Browser sessions are carried by the secure `maxstudio_session` cookie.
servers:
  - url: https://api.maxposty.ru/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /channels:
    get:
      tags: [Channels]
      operationId: listChannels
      summary: List channels owned by the current user
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Channels
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'
        default:
          $ref: '#/components/responses/Problem'
  /channels/discoverable:
    get:
      tags: [Channels]
      operationId: listDiscoverableChannels
      summary: List channels associated with the linked MAX identity
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Discoverable channels
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/DiscoverableChannel'
        default:
          $ref: '#/components/responses/Problem'
  /channels/discoverable/refresh:
    post:
      tags: [Channels]
      operationId: refreshDiscoverableChannels
      summary: Refresh channel inventory from MAX
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverableChannelRefresh'
        default:
          $ref: '#/components/responses/Problem'
...
  /channels/{id}/test:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Channels]
      operationId: testChannel
      summary: Verify the bot permissions for a channel
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
        default:
          $ref: '#/components/responses/Problem'
  /channels/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    delete:
      tags: [Channels]
      operationId: deleteChannel
      summary: Disconnect a channel
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Channel disconnected
        default:
          $ref: '#/components/responses/Problem'
...
  /workspaces/{workspace_id}/channels:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    get:
      tags: [Channels]
      operationId: listWorkspaceChannels
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/ChannelList'}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Channels]
      operationId: createWorkspaceChannelClaim
      summary: Start a verified MAX channel connection for this workspace
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChannelConnectInput'}
      responses:
        '201': {$ref: '#/components/responses/ChannelConnectionClaim'}
        default: {$ref: '#/components/responses/Problem'}
...
  /workspaces/{workspace_id}/channels/{channel_id}:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/ChannelId'}
    get:
      tags: [Channels]
      operationId: getWorkspaceChannel
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/Channel'}
        default: {$ref: '#/components/responses/Problem'}
    patch:
      tags: [Channels]
      operationId: updateWorkspaceChannel
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChannelUpdate'}
      responses:
        '200': {$ref: '#/components/responses/Channel'}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Channels]
      operationId: deleteWorkspaceChannel
      security: [{sessionCookie: []}]
      responses:
        '204': {description: Channel disconnected}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: maxstudio_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    ChannelId:
      name: channel_id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    Problem:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemEnvelope'
...
    Channel:
      description: Workspace channel
      content: {application/json: {schema: {$ref: '#/components/schemas/Channel'}}}
    ChannelList:
      description: Workspace channels
      content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Channel'}}}}
    ChannelConnectionClaim:
      description: Channel ownership verification claim
      content: {application/json: {schema: {$ref: '#/components/schemas/ChannelConnectionClaim'}}}
...
  schemas:
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Problem'
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
...
    ChannelConnectInput:
      type: object
      properties:
        public_link: {type: string, format: uri}
        max_chat_id: {type: string}
        title: {type: string, maxLength: 200}
      anyOf:
        - required: [public_link]
        - required: [max_chat_id]
    ChannelUpdate:
      type: object
      minProperties: 1
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        active: {type: boolean}
...
    Channel:
      type: object
      required:
        - id
        - max_chat_id
        - title
        - participants_count
        - is_channel
        - active
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: string
        max_chat_id:
          type: string
        title:
          type: string
        public_link:
          type: string
          format: uri
        icon_url:
          type: string
        participants_count:
          type: integer
          minimum: 0
        is_channel:
          type: boolean
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DiscoverableChannel:
      type: object
      required: [max_chat_id, title, connected]
      properties:
        max_chat_id:
          type: string
        title:
          type: string
        public_link:
          type: string
          format: uri
        icon_url:
          type: string
        participants_count:
          type: integer
          minimum: 0
        owner_verified:
          type: boolean
        connected:
          type: boolean
        connected_channel_id:
          type: [integer, 'null']
          format: int64
    DiscoverableChannelRefresh:
      type: object
      required: [channels, refreshed, failed]
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/DiscoverableChannel'
        refreshed:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
    ChannelDiagnostics:
      type: object
      required:
        - chat_id
        - type
        - status
        - is_admin
        - permissions
        - can_publish
        - can_edit
        - can_delete
        - missing_required_permissions
      properties:
        chat_id:
          type: string
        type:
          type: string
        status:
          type: string
        is_admin:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        can_publish:
          type: boolean
        can_edit:
          type: boolean
        can_delete:
          type: boolean
        missing_required_permissions:
          type: array
          items:
            type: string
    ChannelConnectionStatus:
      type: string
      enum: [pending, identity_verified, connected, expired, failed]
    ChannelConnectionClaim:
      type: object
      required: [claim_id, status]
      properties:
        claim_id:
          type: string
        status:
          $ref: '#/components/schemas/ChannelConnectionStatus'
        expires_at:
          type: [string, 'null']
          format: date-time
        bot_url:
          type: [string, 'null']
          format: uri
        comparison_code:
          type: [string, 'null']
        requester_label:
          type: [string, 'null']
        bot_name:
          type: [string, 'null']
        bot_username:
          type: [string, 'null']
        message:
          type: [string, 'null']
        channel:
          oneOf:
            - $ref: '#/components/schemas/Channel'
            - type: 'null'
...
    TestResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: string
        title:
          type: string
        channel:
          $ref: '#/components/schemas/Channel'
        diagnostics:
          $ref: '#/components/schemas/ChannelDiagnostics'
...